move copy logic inside ComposeScreen
